refactor: expose action selector
