+++ projects/core/src/select.ts
@@ -1,19 +1,9 @@
 import {inject, KeyValueDiffers, ProviderToken} from "@angular/core";
-import {
-   defer,
-   distinctUntilChanged,
-   filter,
-   map,
-   Observable,
-   shareReplay,
-   startWith,
-   tap
-} from "rxjs";
+import {defer, distinctUntilChanged, filter, map, Observable, shareReplay, startWith} from "rxjs";
 import {ACTION, CONTEXT, EffectScheduler, FLUSHED} from "./providers";
 import {getMeta, selector, setMeta} from "./metadata";
 import {EventType, Metadata, NextEvent, SelectMetadata, StoreEvent} from "./interfaces";
 import {dispatch} from "./dispatch";
-import {untrack} from "./proxy";
 import {events} from "./utils";
 
 export type Selector<T> = {
@@ -60,32 +50,4 @@ export function selectStore<T extends {}>(token: ProviderToken<T>): Observable<T
       setMeta(selector, source, store)
       return source
    }
-}
-
-export function snapshot<T>(source: (...args: any[]) => Observable<T>): T | undefined
-export function snapshot<T>(source: Observable<T>): T | undefined
-export function snapshot<T>(source: Observable<T> | ((...args: any[]) => Observable<T>)): T | undefined {
-   const select = inject(ACTION) as Metadata<SelectMetadata>
-   const { instance } = inject(CONTEXT) as any
-   const scheduler = inject(EffectScheduler)
-   const cacheKey = getMeta(selector, instance) as SelectMetadata
-   const name = "name" in source ? source.name : ""
-   let current = getMeta(cacheKey, instance, select.key) as T
-
-   source = typeof source !== "function"
-      ? source
-      : events(instance).pipe(
-         filter((event: StoreEvent): event is NextEvent => event.name === name && event.type === EventType.Next),
-         map((event) => event.value)
-      )
-
-   dispatch(source, (value) => {
-      current = value
-      setMeta(cacheKey, value, instance, select.key)
-   })
-   const frame = requestAnimationFrame(() => {
-      cancelAnimationFrame(frame)
-      scheduler.dequeue()
-   })
-   return current
 }
